[DX] Extract EventInterface
