+++ src/Akeneo/Bundle/BatchBundle/Event/StepExecutionEvent.php
@@ -3,6 +3,7 @@
 namespace Akeneo\Bundle\BatchBundle\Event;
 
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
+use Akeneo\Component\Batch\Event\EventInterface;
 use Symfony\Component\EventDispatcher\Event;
 
 /**
