+++ src/Akeneo/Bundle/BatchBundle/Tests/Unit/EventListener/SetJobExecutionLogFileSubscriberTest.php
@@ -2,7 +2,7 @@
 
 namespace Akeneo\Bundle\BatchBundle\Tests\Unit\EventListener;
 
-use Akeneo\Bundle\BatchBundle\Event\EventInterface;
+use Akeneo\Component\Batch\Event\EventInterface;
 use Akeneo\Bundle\BatchBundle\EventListener\SetJobExecutionLogFileSubscriber;
 
 /**
