+++ src/Akeneo/Bundle/BatchBundle/Tests/Unit/Step/ItemStepTest.php
@@ -2,7 +2,7 @@
 
 namespace Akeneo\Bundle\BatchBundle\Tests\Unit\Step;
 
-use Akeneo\Bundle\BatchBundle\Event\EventInterface;
+use Akeneo\Component\Batch\Event\EventInterface;
 use Akeneo\Component\Batch\Item\InvalidItemException;
 use Akeneo\Bundle\BatchBundle\Job\BatchStatus;
 use Akeneo\Component\Batch\Step\ItemStep;
