+++ src/Pim/Bundle/BaseConnectorBundle/EventListener/InvalidItemsCollector.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\BaseConnectorBundle\EventListener;
 
-use Akeneo\Bundle\BatchBundle\Event\EventInterface;
+use Akeneo\Component\Batch\Event\EventInterface;
 use Akeneo\Bundle\BatchBundle\Event\InvalidItemEvent;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
