+++ src/Pim/Bundle/BaseConnectorBundle/EventListener/JobExecutionArchivist.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\BaseConnectorBundle\EventListener;
 
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
-use Akeneo\Bundle\BatchBundle\Event\EventInterface;
+use Akeneo\Component\Batch\Event\EventInterface;
 use Akeneo\Bundle\BatchBundle\Event\JobExecutionEvent;
 use Pim\Bundle\BaseConnectorBundle\Archiver\ArchiverInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
