+++ src/Pim/Bundle/VersioningBundle/EventSubscriber/AddContextSubscriber.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\VersioningBundle\EventSubscriber;
 
 use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
-use Akeneo\Bundle\BatchBundle\Event\EventInterface;
+use Akeneo\Component\Batch\Event\EventInterface;
 use Akeneo\Bundle\BatchBundle\Event\JobExecutionEvent;
 use Pim\Bundle\VersioningBundle\Manager\VersionContext;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
